+++ src/prefect/cli/__init__.py
@@ -7,6 +7,7 @@ import prefect
 
 from .agent import agent as _agent
 from .auth import auth as _auth
+from .create import create as _create
 from .describe import describe as _describe
 from .execute import execute as _execute
 from .get import get as _get
@@ -31,10 +32,11 @@ def cli():
         describe    Retrieve detailed object descriptions
 
     \b
-    Execution Commands:
+    Action Commands:
+        agent       Manage agents
+        create      Create objects
         execute     Execute a flow's environment
         run         Run a flow
-        agent       Manage agents
 
     \b
     Setup Commands:
@@ -50,6 +52,7 @@ def cli():
 
 cli.add_command(_agent)
 cli.add_command(_auth)
+cli.add_command(_create)
 cli.add_command(_describe)
 cli.add_command(_execute)
 cli.add_command(_get)